AI, Text mode, OWM position,
Added a text mode to use the Biscotte in the terminal (easy debug and testing). Added AI when no commande is detected after the keyword. Added ip based position for OWM. Made the website and program scanning stop after one is detected. Change names in Start.bat. Removed the toast for the reminders. Added rich print and prompt